project/backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
Flask-CORS
python-dotenv
firebase-admin
scikit-learn
numpy
requests
apscheduler
pdfplumber
PyPDF2
gunicorn
